Clear Last Data from Signal
lastDataFired is a strong reference to the last object that was sent through the Signal. If sending objects with a wide reference graph, this can lead to some unexpected memory leaks.

Ideally the lastDataFired should be a weak reference, but that is not possible with Swift non-class types.

This can already be achieved by setting fireCount to 0 and the lastDataFired to nil, but an explicit API call to do some clean up is probably a better design.